fixed the israeli-palestinian conflict

+++ src/main/java/org/marinemc/util/MathUtils.java
@@ -1,7 +1,25 @@
+///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+// MarineStandalone is a minecraft server software and API.
+// Copyright (C) MarineMC (marinemc.org)
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+///////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 package org.marinemc.util;
 
 import org.marinemc.util.operations.ArgumentedOperation;
-import org.marinemc.util.vectors.Vector2d;
 import org.marinemc.util.vectors.Vector2i;
 
 public class MathUtils {
@@ -28,21 +46,21 @@ public class MathUtils {
 
 		return (w < x || w > X) && (h < y || h > Y);
 	}
-	
-	public static void circleOperation(final ArgumentedOperation<Vector2d> op, int x, int y, int radius, int step) {
+
+	public static void circleOperation(final ArgumentedOperation<Vector2i> op, int x, int y, int radius, int step) {
 		int theta = 0;
 		
 		while(theta <= 360) {
-			op.action(new Vector2d(x + radius*Math.cos(theta), y + radius*Math.sin(theta)));
+			op.action(new Vector2i((int) (x + radius * Math.cos(theta)), (int) (y + radius * Math.sin(theta))));
 			theta += step;
 		}
 	}
-	
-	public static void circleOperation(final ArgumentedOperation<Vector2d> op, int x, int y, int radius) {
+
+	public static void circleOperation(final ArgumentedOperation<Vector2i> op, int x, int y, int radius) {
 		int theta = 0;
 		
 		while(theta++ <= 360)
-			op.action(new Vector2d(x + radius*Math.cos(theta), y + radius*Math.sin(theta)));
+			op.action(new Vector2i((int) (x + radius * Math.cos(theta)), (int) (y + radius * Math.sin(theta))));
 	}
 	
 	public static void rectOperation(final ArgumentedOperation<Vector2i> op, final int X, final int Y, final int width, final int height) {
